refactor: new file structure in `js` websocket client (#1609)
